requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-dotenv
email-validator
pymongo<4.7
